+++ src/tests/formatter_test.py
@@ -1,6 +1,6 @@
 import unittest
 
-from src.utility.formatters import format_degrees_to_coordinate_lat, format_degrees_to_coordinate_long, format_minutes, \
+from src.converters.utility import format_degrees_to_coordinate_lat, format_degrees_to_coordinate_long, format_minutes,\
     format_degrees_long, format_degrees_lat
