+++ src/tests/importer_test.py
@@ -1,57 +1,78 @@
 import unittest
-import pandas as pd
-from numpy import dtype
+from io import StringIO
 
-from src.importers.text_importer import rename_duplicate_columns
+import pandas as pd
+
+from src.importers.text_importer import rename_duplicate_columns, transpose_df, import_file, TextImporter
 
 import numpy as np
 
-from src.modules.import_module import clean_datetime_columns, clean_scientific_columns
-
 
 class TextImporterTests(unittest.TestCase):
-    df = pd.DataFrame(
-        np.array([['19', '10', '10', '10', '10', '10']]),
-        columns=['EVENT HEADER - TIME (YY)',
-                 'EVENT HEADER - TIME (MM)',
-                 'EVENT HEADER - TIME (DD)',
-                 'EVENT HEADER - TIME (HH)',
-                 'EVENT HEADER - TIME (MM)',
-                 'EVENT HEADER - TIME (SS)'])
+    csv = StringIO("""
+        --
+        --
+        EVENT NUMBER: 1
+        1_COL1: VALUE1
+        1_COL2: VALUE2
+        --
+        --
+        EVENT NUMBER: 2
+        2_COL1: VALUE1
+        2_COL2: VALUE2
+        --
+        --
+        """)
 
     def test_clean_duplicate_columns(self):
         """
         Should rename duplicate columns with + '.i'
         """
 
-        df = rename_duplicate_columns(self.df.copy())
+        df = pd.DataFrame(
+            np.array([['19', '10', '10', '10', '10', '10']]),
+            columns=['EVENT HEADER - TIME (YY)',
+                     'EVENT HEADER - TIME (MM)',
+                     'EVENT HEADER - TIME (DD)',
+                     'EVENT HEADER - TIME (HH)',
+                     'EVENT HEADER - TIME (MM)',
+                     'EVENT HEADER - TIME (SS)'])
 
-        self.assertEqual(['EVENT HEADER - TIME (YY)', 'EVENT HEADER - TIME (MM)', 'EVENT HEADER - TIME (DD)',
-                          'EVENT HEADER - TIME (HH)', 'EVENT HEADER - TIME (MM).1', 'EVENT HEADER - TIME (SS)'],
-                         list(df.columns))
+        df = rename_duplicate_columns(df.copy())
 
-    def test_clean_datetime_columns(self):
-        df = rename_duplicate_columns(self.df.copy())
-        df = clean_datetime_columns(df)
+        actual = list(df.columns)
+        expect = ['EVENT HEADER - TIME (YY)', 'EVENT HEADER - TIME (MM)', 'EVENT HEADER - TIME (DD)',
+                  'EVENT HEADER - TIME (HH)', 'EVENT HEADER - TIME (MM).1', 'EVENT HEADER - TIME (SS)']
 
-        self.assertEqual(['DATE', 'TIME'], list(df.columns))
+        self.assertEqual(expect, actual)
 
-    def test_clean_scientific_columns(self):
+    def test_transpose_df(self):
+        """
+        Should create a DF where every event is a row
+        """
+
         df = pd.DataFrame({
-            'col1': ['1.364571e+001'],
-            'col2': ['-2.485793e+002'],
-            'col3': ['test'],
-            'col4': ['test']
+            'NAME': ['COL1', 'COL2', 'COL3', 'COL1', 'COL2', 'COL3'],
+            'VALUE': ['VALUE1', 'VALUE2', 'VALUE3', 'VALUE1', 'VALUE2', 'VALUE3'],
+            'EVENT NUMBER': [1, 1, 1, 2, 2, 2]
         })
 
-        for col in df.dtypes:
-            with self.subTest(col=col):
-                self.assertEqual(dtype('object'), col)
+        actual = transpose_df(df).values.tolist()
+        expect = [['COL1', 'COL2', 'COL3', np.nan, np.nan, np.nan],
+                  ['VALUE1', 'VALUE2', 'VALUE3', np.nan, np.nan, np.nan],
+                  [np.nan, np.nan, np.nan, 'COL1', 'COL2', 'COL3'],
+                  [np.nan, np.nan, np.nan, 'VALUE1', 'VALUE2', 'VALUE3']]
 
-        df = clean_scientific_columns(df)
+        self.assertEqual(expect, actual)
 
-        asserts = [dtype('float64'), dtype('float64'), dtype('object'), dtype('object')]
+    def test_import_file(self):
+        actual = import_file(self.csv).reset_index().values.tolist()
 
-        for i, col in enumerate(df.dtypes):
-            with self.subTest(col=col):
-                self.assertEqual(asserts[i], col)
+        expect = [['EVENT NUMBER', '1', 1],
+                  ['1_COL1', 'VALUE1', 1],
+                  ['1_COL2', 'VALUE2', 1],
+                  ['EVENT NUMBER', '2', 2],
+                  ['2_COL1', 'VALUE1', 2],
+                  ['2_COL2', 'VALUE2', 2]]
+
+        self.assertEqual(expect, actual)
